+++ BNA_KC_Vehicles/Light/ATRT/CfgFunctions.hpp
@@ -1,0 +1,14 @@
+class CfgFunctions
+{
+    class BNAKC
+    {
+        class ATRT
+        {
+            file = "\BNA_KC_Vehicles\Light\ATRT\Data\Functions";
+            class initATRT {};
+            class mountATRT {};
+            class dismountATRT {};
+            class spawnATRTSmoke {};
+        };
+    };
+};
