+++ BNA_KC_Vehicles/Light/ATRT/Config.cpp
@@ -1,4 +1,5 @@
 #include "CfgPatches.hpp"
+#include "CfgFunctions.hpp"
 #include "..\..\Common\Macros.hpp"
