+++ BNA_KC_Vehicles/Light/ATRT/Data/Functions/fn_dismountATRT.sqf
