+++ BNA_KC_Vehicles/Light/ATRT/Data/Functions/fn_initATRT.sqf
