+++ BNA_KC_Vehicles/Light/ATRT/Data/Functions/fn_spawnATRTSmoke.sqf
